requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
bcrypt
pyjwt
passlib[bcrypt]  # ✅ Add this for password hashing
